requirements.txt
flask>=3.0.0
openpyxl>=3.1.0
qrcode>=7.4.0
Pillow>=10.0.0
